improvements to pyana run interface. bugfix to pyana_image_beta
git-svn-id: https://pswww.slac.stanford.edu/svn-readonly/psdmrepo@2520 b967ad99-d558-0410-b138-e0f6c56caec7

+++ XtcExplorer/trunk/src/XtcPyanaControl.py
@@ -61,19 +61,27 @@ from pyana import pyanamod
 #---------------------
 class myPopen(subprocess.Popen):
     def kill(self, signal = signal.SIGTERM):
-        os.kill(self.pid, signal)
-        print "pyana process %d has been killed "% self.pid
+        code = self.poll()
+        if code is None: 
+            os.kill(self.pid, signal)
+            print "pyana process %d has been killed "% self.pid
+            return 1
+        else :
+            print "No pyana process to kill... "
+            if code==0 : 
+                print "Finished successfully"
+            else :
+                print "Exited with code ", code
+                code = 1
+            return code
 
 #class MyThread( threading.Thread ):
 class MyThread( QtCore.QThread ):
-    """Run pyana module in a separate thread.
-    This allows the GUI windows to stay active.
-    The only problem is that Matplotlib windows
-    need to me made beforehand, by the Gui. Not
-    in pyana. Not a problem as long as they are
-    declared beforehand. Pyana can still call the
-    plt.figure command, that way it can be run
-    standalone or from the GUI. 
+    """Run pyana module in a separate thread. This allows the GUI windows
+    to stay active. The only problem is that Matplotlib windows need to me
+    made beforehand, by the GUI -- not in pyana. Not a problem as long as
+    they are declared beforehand. Pyana can still call the plt.figure command,
+    that way it can be run standalone or from the GUI. 
     In principle...
     Still some issues to look into:
     - matplotlib figure must be created before pyana runs
@@ -97,7 +105,13 @@ class MyThread( QtCore.QThread ):
         print " !   python threads cannot be interupted..."
         print " !   you'll have to wait..."
         print " !   or ^Z and kill the whole xtcbrowser process."
-
+        # SOLUTION: Run the Plot gui in a subprocess, this subprocess then calls the thread.
+        # That should keep the other GUIs active, while the Plot GUI waits (or not) for pyana.
+        # Killing pyana thread then requires killing the Plot GUI subprocess.
+        #self.terminate()  ... hangs indefinitely & freezes up the GUI
+        #self.exit(0) ..... does nothing
+        #self.quit() .... does nothing
+        print "done killing"
     
 
 class XtcPyanaControl ( QtGui.QWidget ) :
@@ -147,6 +161,7 @@ class XtcPyanaControl ( QtGui.QWidget ) :
         self.checkboxes = []
 
         self.proc_pyana = None
+        self.proc_status = None
         self.configfile = None
 
         self.pvWindow = None
@@ -1027,9 +1042,16 @@ Start with selecting data of interest to you from list on the left and general r
 
         if self.pyana_button is None: 
             self.pyana_button = QtGui.QPushButton("&Run pyana")
+            self.pyana_button.setMaximumWidth(120)
+            self.proc_status = QtGui.QLabel("")
+            
             self.connect(self.pyana_button, QtCore.SIGNAL('clicked()'), self.run_pyana)
-            self.layout.addWidget( self.pyana_button )
-            self.layout.setAlignment( self.pyana_button, QtCore.Qt.AlignRight )
+
+            pyana_button_line = QtGui.QHBoxLayout()
+            pyana_button_line.addWidget( self.proc_status )
+            pyana_button_line.addWidget( self.pyana_button )
+            self.layout.addLayout( pyana_button_line  )
+            self.layout.setAlignment( pyana_button_line, QtCore.Qt.AlignRight )
         else :
             self.pyana_button.show()
 
@@ -1047,8 +1069,6 @@ Start with selecting data of interest to you from list on the left and general r
             print "set the EDITOR variable in your shell environment."
             proc_emacs = myPopen("emacs %s" % self.configfile, shell=True)
 
-        stdout_value = proc_emacs.communicate()[0]
-        print stdout_value
         #proc_emacs = MyThread("emacs %s" % self.configfile) 
         #proc_emacs.start()
         
@@ -1124,8 +1144,10 @@ Start with selecting data of interest to you from list on the left and general r
         if 1 :
             # calling a new process
             self.proc_pyana = myPopen(lpoptions) # this runs in separate thread.
-            #stdout_value = proc_pyana.communicate()[0]
-            #print stdout_value
+            self.proc_status.setText("pyana process %d is running "%self.proc_pyana.pid)
+            # this blocks: 
+            #stdout_value = self.proc_pyana.communicate()[0]
+            #print "Here's what the pyana process communicates: ",stdout_value
             # the benefit of this option is that the GUI remains unlocked
             # the drawback is that pyana needs to supply its own plots, ie. no Qt plots?
             
@@ -1148,10 +1170,13 @@ Start with selecting data of interest to you from list on the left and general r
             self.proc_pyana = MyThread(lpoptions)
             self.proc_pyana.start()
             print "I'm back"
-            
+
+
+        self.pyana_button.setDisabled(True)
             
         if self.quit_pyana_button is None :
             self.quit_pyana_button = QtGui.QPushButton("&Quit pyana")
+            self.quit_pyana_button.setMaximumWidth(120)
             self.connect(self.quit_pyana_button, QtCore.SIGNAL('clicked()'), self.quit_pyana )
             self.layout.addWidget( self.quit_pyana_button )
             self.layout.setAlignment( self.quit_pyana_button, QtCore.Qt.AlignRight )
@@ -1162,7 +1187,14 @@ Start with selecting data of interest to you from list on the left and general r
         """Kill the pyana process
         """
         if self.proc_pyana :
-            self.proc_pyana.kill()
+
+            statustext = {1 : "process %d killed"%self.proc_pyana.pid,
+                          0 : "process %d finished successfully"%self.proc_pyana.pid }
+            status = self.proc_pyana.kill()
+
+            self.pyana_button.setDisabled(False)        
+            self.proc_status.setText(statustext[status])
+            self.quit_pyana_button.hide()
             return
 
         print "No pyana process to stop"
